refactor: フェーズ 2 レビュー指摘事項を対応
- makeEvents エイリアスを domain 層から削除、shim (game-logic.ts) にのみ残す
- math-utils.shuffle に randomFn 引数を追加、quiz 側の shuffleArray を統合削除
- pickQuestion を options オブジェクト化（後方互換: Set<number> も受付可）
- buildStudyPool を options オブジェクト化（後方互換: number も受付可）
- achievement-types.ts から SavedGameResult 等を game-result-types.ts に分離
- コメント修正: 実績定義「15個」→「20個」
