+++ .docs/aqs-20260315-01/tasks.md
@@ -109,7 +109,7 @@
 ### 2-8. フェーズ 2 完了確認
 - [x] `npm run ci` パス
 - [x] テストパス確認（既存テスト33スイート598テスト全パス）
-- [ ] コミット作成
+- [x] コミット作成（022e22f）
 
 ---
 
@@ -453,7 +453,7 @@
 |---------|---------|--------|------|
 | 0: 準備 | 10 | 10 | **完了** |
 | 1: 型定義分割 | 13 | 10 | **完了**（レビュー指摘3件は後続フェーズで対応） |
-| 2: ドメイン層抽出 | 39 | 38 | **完了**（コミット待ち） |
+| 2: ドメイン層抽出 | 39 | 39 | **完了** |
 | 3: DbC 導入 | 19 | 0 | 未着手 |
 | 4: インフラ層分離 | 25 | 0 | 未着手 |
 | 5: アプリケーション層 | 11 | 0 | 未着手 |
